refactor :: PlatformType enum 패키지 이동
