+++ core/common/src/main/java/team/retum/common/enums/PlatformType.kt
@@ -1,4 +1,4 @@
-package team.retum.common
+package team.retum.common.enums
 
 enum class PlatformType {
     WEB,
